Move app component to src/components
